pyproject.toml
[project]
name = "django-monolith"
version = "0.1.0"
description = "Default django monolith project"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "gunicorn>=25.1.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
]
